requirements.txt
Django==1.10.1
djangorestframework==3.5.3
django-cors-middleware==1.3.1
Markdown==2.6.7
django-filter==1.0.1
iso8601==0.1.11
promise==0.4.2
singledispatch==3.4.0.3
six==1.10.0
